Change precision to double

+++ include/boundaryConditionTypeInterface.hpp
@@ -6,7 +6,7 @@ depending on the user choice*/
 
 class BoundaryConditionTypeInterface{
     public:
-           virtual void updateBoundaries(float* _ptr2SolutionInitial, int _Nx=0, int _Ny=0, int _Nz=0, float _surfaceTemperature=0.)=0;
+           virtual void updateBoundaries(double* _ptr2SolutionInitial, int _Nx=0, int _Ny=0, int _Nz=0, double _surfaceTemperature=0.)=0;
            virtual ~BoundaryConditionTypeInterface() = default;
 };
