+++ include/boundaryCreatorInterface.hpp
@@ -12,7 +12,7 @@ The interface of this class interacts with the main*/
 class BoundaryCreatorInterface{
     public:
            //CTOR
-           BoundaryCreatorInterface(float* _ptr2State, int _Nx, int _Ny, int _Nz=0, float _surfaceTemp=0.): m_ptr2State(_ptr2State),
+           BoundaryCreatorInterface(double* _ptr2State, int _Nx, int _Ny, int _Nz=0, double _surfaceTemp=0.): m_ptr2State(_ptr2State),
            m_Nx(_Nx), m_Ny(_Ny), m_Nz(_Nz), m_surfaceTemp(_surfaceTemp){}
            //Method to set the boundary condition strategy
            void setBoundaryConditionType(){
@@ -28,13 +28,13 @@ class BoundaryCreatorInterface{
            virtual ~BoundaryCreatorInterface() = default;
     protected:
            //Member to store a pointer to the state whose boundary has to be updated
-           float* m_ptr2State;
+           double* m_ptr2State;
            //Member to store the grid size in 3 directions
            int m_Nx, m_Ny, m_Nz;
            //TODO Get rid of the need to store the surface temp here - it only belongs to the
            //dirichlet class?
            //Member to store the surface temperature
-           float m_surfaceTemp;
+           double m_surfaceTemp;
            //Member to store the type of boundary condition that must be used.
            // The factory method must return the updateBoundaries() of this type
            BoundaryConditionTypeInterface* m_BCType;
